Address PR review: fix NSActivity comment, nullable param, named constant, int counter
- M1: Fix misleading comment — 0x00FFFFFF is NSActivityUserInitiated, not
  AllowingIdleSystemSleep (0x00EFFFFF). Kept the value since we want full
  App Nap prevention for WebSocket bridge reliability.
- N1: Fix CS8765 — NSDictionary? launchOptions nullable to match base.
- N2: Extract magic number 10 to InitialHistoryLoadThreshold constant.
- N3: Upgrade generation counter from byte to int to eliminate theoretical
  wrap-around at 255 sub-turns.

+++ PolyPilot/Services/CopilotService.cs
@@ -17,7 +17,7 @@ public partial class CopilotService : IAsyncDisposable
     // before the server processes the close and removes them from its broadcast
     private readonly ConcurrentDictionary<string, byte> _recentlyClosedRemoteSessions = new();
     // Sessions currently receiving streaming content via bridge events — history sync skipped to avoid duplicates
-    private readonly ConcurrentDictionary<string, byte> _remoteStreamingSessions = new();
+    private readonly ConcurrentDictionary<string, int> _remoteStreamingSessions = new();
 
     /// <summary>
     /// Whether a session's history is still being synced after a turn completed (streaming guard active).
